+++ test/index_test.rb
@@ -40,6 +40,8 @@ class IndexTest < Minitest::Test
     assert_equal "L2", index.distance_type
     assert_equal 10, index.edge_size_for_creation
     assert_equal 40, index.edge_size_for_search
+    assert_equal :float, index.object_type
+
     assert_equal [1, 2, 3], index.batch_insert(objects)
     path = Dir.mktmpdir
     index.save(path)